// src/models/UserMiror.interface.ts
import { UserAddressLocation } from "./UserAddressLocation";

export class UsersMirror {
      $key: string;
      Password: string;
      Email: string;
      IsFollowed: boolean;
      Age: number;
      UserImg: string;
      UserLocation: UserAddressLocation;
      MyPosts: [string];
      userid:string;

}